+++ cnx/src/bar.rs
@@ -50,6 +50,8 @@ fn create_surface(
     screen_idx: usize,
     window_id: u32,
     height: u16,
+    width: Option<u16>,
+    offset: Offset,
 ) -> Result<(u16, cairo::XCBSurface)> {
     let screen = conn
         .get_setup()
@@ -61,15 +63,15 @@ fn create_surface(
         (xcb::CW_EVENT_MASK, xcb::EVENT_MASK_EXPOSURE),
     ];
 
-    let width = screen.width_in_pixels();
+    let width = width.unwrap_or_else(|| screen.width_in_pixels());
 
     xcb::create_window(
         conn,
         xcb::COPY_FROM_PARENT as u8,
         window_id,
         screen.root(),
-        0,
-        0,
+        offset.x,
+        offset.y,
         width,
         height,
         0,
@@ -110,6 +112,13 @@ pub enum Position {
     Bottom,
 }
 
+/// A struct specifying the `x` and `y` offset
+#[derive(Default, Clone, Copy)]
+pub struct Offset {
+    pub x: i16,
+    pub y: i16,
+}
+
 pub struct Bar {
     position: Position,
 
@@ -120,12 +129,13 @@ pub struct Bar {
     surface: cairo::XCBSurface,
     width: u16,
     height: u16,
+    offset: Offset,
 
     contents: Vec<Vec<ComputedText>>,
 }
 
 impl Bar {
-    pub fn new(position: Position) -> Result<Bar> {
+    pub fn new(position: Position, width: Option<u16>, offset: Offset) -> Result<Bar> {
         let (conn, screen_idx) =
             xcb::Connection::connect(None).context("Failed to connect to X server")?;
         let screen_idx = screen_idx as usize;
@@ -135,7 +145,7 @@ impl Bar {
         // our window once we know how big it needs to be. However, it seems to need
         // to be bigger than 0px, or either Xcb/Cairo (or maybe QTile?) gets upset.
         let height = 1;
-        let (width, surface) = create_surface(&conn, screen_idx, window_id, height)?;
+        let (width, surface) = create_surface(&conn, screen_idx, window_id, height, width, offset)?;
 
         let ewmh_conn = ewmh::Connection::connect(conn)
             .map_err(|(e, _)| e)
@@ -148,6 +158,7 @@ impl Bar {
             surface,
             width,
             height,
+            offset,
             position,
             contents: Vec::new(),
         };
@@ -217,8 +228,11 @@ impl Bar {
             // If we're at the bottom of the screen, we'll need to update the
             // position of the window.
             let y = match self.position {
-                Position::Top => 0,
-                Position::Bottom => self.screen()?.height_in_pixels() - self.height,
+                Position::Top => self.offset.y.max(0) as u16,
+                Position::Bottom => {
+                    let h = (self.screen()?.height_in_pixels() - self.height) as i32;
+                    h.checked_add(self.offset.y as i32).unwrap_or(h).max(0) as u16
+                }
             };
 
             // Update the height/position of the XCB window and the height of the Cairo surface.